feat: 精算フローの共通化 - ログイン状態に関係なく同じUIとロジックを使用
